+++ sources/server/changelog.md
@@ -1,5 +1,6 @@
 ### 2.4.0-beta.1
 * New feature #575 : add keyword without history
+* New feature (plugin API) : web requests done by plugins now include proxy settings from Yadoms. Just need to rebuild plugin.
 
 
 ### 2.3.0